requirements.txt
google-genai>=1.17.0
marker-pdf
chromadb
sentence-transformers
pypdf